+++ TerminalGameEngine/AutomaticDoor.cpp
@@ -20,8 +20,8 @@ AutomaticDoor::AutomaticDoor
 (
     int xPos,
     int yPos,
-    uint sizeX,
-    uint sizeY,
+    size_t sizeX,
+    size_t sizeY,
     char modelChar
 ) : VariableSizeCollider(xPos, yPos, sizeX, sizeY, modelChar)
 {
