+++ TerminalGameEngine/StaticCollider.cpp
@@ -4,8 +4,8 @@ StaticCollider::StaticCollider
 (
     int xPos, 
     int yPos, 
-    uint sizeX,
-    uint sizeY,
+    size_t sizeX,
+    size_t sizeY,
     char modelChar
 ) : VariableSizeCollider(xPos, yPos, sizeX, sizeY, modelChar)
 {
